it works
The error message "Invalid stream - missing peerId property" indicates that the peerId property is missing from the stream object, which is preventing the video element from being added to the grid.

To fix this error, you need to ensure that the peerId property is correctly set on the stream object before passing it to the addVideoStream() function.

In the code that initializes the getUserMedia stream, you should modify the then() callback to set the peerId property on the stream object before calling addVideoStream(). Here's an example of how you could do this: